debug/картинки

+++ src/main/java/ru/example/bookmarket/service/ImageService.java
@@ -3,18 +3,20 @@ package ru.example.bookmarket.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+import ru.example.bookmarket.exception.ImageNotFoundException;
 import ru.example.bookmarket.model.Image;
 import ru.example.bookmarket.repository.ImageRepository;
 import ru.example.bookmarket.util.Converter;
 
 import javax.persistence.EntityNotFoundException;
+import javax.transaction.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class ImageService {
-    ImageRepository imageRepository;
+    private final ImageRepository imageRepository;
 
     public List<Image> save(List<MultipartFile> files) {
         List<Image> images = files.stream()
@@ -23,16 +25,17 @@ public class ImageService {
         return imageRepository.saveAll(images);
     }
 
+    @Transactional
     public Image findById(Long id) {
         return imageRepository.findById(id)
                 .orElseThrow(EntityNotFoundException::new);
     }
 
     public void deleteById(Long id) {
-        try {
+        if (imageRepository.existsById(id)) {
             imageRepository.deleteById(id);
-        } catch (EntityNotFoundException e) {
-            throw new EntityNotFoundException("Image with id '%d' not exist");
-        }
+        } else throw new ImageNotFoundException(id);
+
+
     }
 }
